+++ src/ODE.jl
@@ -515,155 +515,3 @@ function Base.show(io::IO, ode::ODE)
 end
 
 #------------------------------------------------------------------------------
-"""
-    function mtk_to_si(de::ModelingToolkit.AbstractTimeDependentSystem, measured_quantities::Array{ModelingToolkit.Equation})
-    function mtk_to_si(de::ModelingToolkit.AbstractTimeDependentSystem, measured_quantities::Array{SymbolicUtils.BasicSymbolic})
-
-Input:
-- `de` - ModelingToolkit.AbstractTimeDependentSystem, a system for identifiability query
-- `measured_quantities` - array of output functions (as equations of just functions)
-
-Output:
-- `ODE` object containing required data for identifiability assessment
-- `conversion` dictionary from the symbols in the input MTK model to the variable
-  involved in the produced `ODE` object
-"""
-function mtk_to_si(
-    de::ModelingToolkit.AbstractTimeDependentSystem,
-    measured_quantities::Array{ModelingToolkit.Equation},
-)
-    return __mtk_to_si(
-        de,
-        [(replace(string(e.lhs), "(t)" => ""), e.rhs) for e in measured_quantities],
-    )
-end
-
-function mtk_to_si(
-    de::ModelingToolkit.AbstractTimeDependentSystem,
-    measured_quantities::Array{<:Symbolics.Num},
-)
-    return __mtk_to_si(
-        de,
-        [("y$i", Symbolics.value(e)) for (i, e) in enumerate(measured_quantities)],
-    )
-end
-
-function mtk_to_si(
-    de::ModelingToolkit.AbstractTimeDependentSystem,
-    measured_quantities::Array{<:SymbolicUtils.BasicSymbolic},
-)
-    return __mtk_to_si(de, [("y$i", e) for (i, e) in enumerate(measured_quantities)])
-end
-
-#------------------------------------------------------------------------------
-# old name kept for compatibility purposes
-
-function preprocess_ode(
-    de::ModelingToolkit.AbstractTimeDependentSystem,
-    measured_quantities::Array{ModelingToolkit.Equation},
-)
-    @warn "Function `preprocess_ode` has been renamed to `mtk_to_si`. The old name can be still used but will disappear in the future releases."
-    return mtk_to_si(de, measured_quantities)
-end
-
-function preprocess_ode(
-    de::ModelingToolkit.AbstractTimeDependentSystem,
-    measured_quantities::Array{<:Symbolics.Num},
-)
-    @warn "Function `preprocess_ode` has been renamed to `mtk_to_si`. The old name can be still used but will disappear in the future releases."
-    return mtk_to_si(de, measured_quantities)
-end
-
-function preprocess_ode(
-    de::ModelingToolkit.AbstractTimeDependentSystem,
-    measured_quantities::Array{<:SymbolicUtils.BasicSymbolic},
-)
-    @warn "Function `preprocess_ode` has been renamed to `mtk_to_si`. The old name can be still used but will disappear in the future releases."
-    return mtk_to_si(de, measured_quantities)
-end
-
-#------------------------------------------------------------------------------
-"""
-    function __mtk_to_si(de::ModelingToolkit.AbstractTimeDependentSystem, measured_quantities::Array{Tuple{String, SymbolicUtils.BasicSymbolic}})
-
-Input:
-- `de` - ModelingToolkit.AbstractTimeDependentSystem, a system for identifiability query
-- `measured_quantities` - array of input function in the form (name, expression)
-
-Output:
-- `ODE` object containing required data for identifiability assessment
-- `conversion` dictionary from the symbols in the input MTK model to the variable
-  involved in the produced `ODE` object
-"""
-function __mtk_to_si(
-    de::ModelingToolkit.AbstractTimeDependentSystem,
-    measured_quantities::Array{<:Tuple{String, <:SymbolicUtils.BasicSymbolic}},
-)
-    polytype = StructuralIdentifiability.Nemo.fmpq_mpoly
-    fractype = StructuralIdentifiability.Nemo.Generic.Frac{polytype}
-    diff_eqs =
-        filter(eq -> !(ModelingToolkit.isoutput(eq.lhs)), ModelingToolkit.equations(de))
-    # performing full structural simplification
-    if length(observed(de)) > 0
-        rules = Dict(s.lhs => s.rhs for s in observed(de))
-        while any([
-            length(intersect(get_variables(r), keys(rules))) > 0 for r in values(rules)
-        ])
-            rules = Dict(k => SymbolicUtils.substitute(v, rules) for (k, v) in rules)
-        end
-        diff_eqs = [SymbolicUtils.substitute(eq, rules) for eq in diff_eqs]
-    end
-
-    y_functions = [each[2] for each in measured_quantities]
-    inputs = filter(v -> ModelingToolkit.isinput(v), ModelingToolkit.states(de))
-    state_vars = filter(
-        s -> !(ModelingToolkit.isinput(s) || ModelingToolkit.isoutput(s)),
-        ModelingToolkit.states(de),
-    )
-    params = ModelingToolkit.parameters(de)
-    t = ModelingToolkit.arguments(diff_eqs[1].lhs)[1]
-    params_from_measured_quantities = union(
-        [filter(s -> !istree(s), get_variables(y[2])) for y in measured_quantities]...,
-    )
-    params = union(params, params_from_measured_quantities)
-
-    input_symbols = vcat(state_vars, inputs, params)
-    generators = vcat(string.(input_symbols), [e[1] for e in measured_quantities])
-    generators = map(g -> replace(g, "(t)" => ""), generators)
-    R, gens_ = Nemo.PolynomialRing(Nemo.QQ, generators)
-    y_vars = Vector{polytype}([str_to_var(e[1], R) for e in measured_quantities])
-    symb2gens = Dict(input_symbols .=> gens_[1:length(input_symbols)])
-
-    x_vars = Vector{polytype}()
-
-    state_eqn_dict = Dict{polytype, Union{polytype, fractype}}()
-    out_eqn_dict = Dict{polytype, Union{polytype, fractype}}()
-
-    for i in 1:length(diff_eqs)
-        x = substitute(state_vars[i], symb2gens)
-        push!(x_vars, x)
-        if !(diff_eqs[i].rhs isa Number)
-            state_eqn_dict[x] = eval_at_nemo(diff_eqs[i].rhs, symb2gens)
-        else
-            state_eqn_dict[x] = R(diff_eqs[i].rhs)
-        end
-    end
-    for i in 1:length(measured_quantities)
-        out_eqn_dict[y_vars[i]] = eval_at_nemo(measured_quantities[i][2], symb2gens)
-    end
-
-    inputs_ = [substitute(each, symb2gens) for each in inputs]
-    if isequal(length(inputs_), 0)
-        inputs_ = Vector{polytype}()
-    end
-    return (
-        StructuralIdentifiability.ODE{polytype}(
-            x_vars,
-            y_vars,
-            state_eqn_dict,
-            out_eqn_dict,
-            inputs_,
-        ),
-        symb2gens,
-    )
-end
