+++ src/discrete.jl
@@ -286,105 +286,3 @@ function _assess_local_identifiability_discrete_aux(
 end
 
 # ------------------------------------------------------------------------------
-
-"""
-    function assess_local_identifiability(
-        dds::ModelingToolkit.DiscreteSystem; 
-        measured_quantities=Array{ModelingToolkit.Equation}[], 
-        funcs_to_check=Array{}[], 
-        known_ic=Array{}[],
-        prob_threshold::Float64=0.99)
-
-Input:
-- `dds` - the DiscreteSystem object from ModelingToolkit (with **difference** operator in the right-hand side)
-- `measured_quantities` - the measurable outputs of the model
-- `funcs_to_check` - functions of parameters for which to check identifiability (all parameters and states if not specified)
-- `known_ic` - functions (of states and parameter) whose initial conditions are assumed to be known
-- `prob_threshold` - probability of correctness
-
-Output:
-- the result is an (ordered) dictionary from each function to to boolean;
-
-The result is correct with probability at least `prob_threshold`.
-"""
-function assess_local_identifiability(
-    dds::ModelingToolkit.DiscreteSystem;
-    measured_quantities = Array{ModelingToolkit.Equation}[],
-    funcs_to_check = Array{}[],
-    known_ic = Array{}[],
-    prob_threshold::Float64 = 0.99,
-    loglevel = Logging.Info,
-)
-    restart_logging(loglevel = loglevel)
-    with_logger(_si_logger[]) do
-        return _assess_local_identifiability(
-            dds,
-            measured_quantities = measured_quantities,
-            funcs_to_check = funcs_to_check,
-            known_ic = known_ic,
-            prob_threshold = prob_threshold,
-        )
-    end
-end
-
-function _assess_local_identifiability(
-    dds::ModelingToolkit.DiscreteSystem;
-    measured_quantities = Array{ModelingToolkit.Equation}[],
-    funcs_to_check = Array{}[],
-    known_ic = Array{}[],
-    prob_threshold::Float64 = 0.99,
-)
-    if length(measured_quantities) == 0
-        if any(ModelingToolkit.isoutput(eq.lhs) for eq in ModelingToolkit.equations(dds))
-            @info "Measured quantities are not provided, trying to find the outputs in input dynamical system."
-            measured_quantities = filter(
-                eq -> (ModelingToolkit.isoutput(eq.lhs)),
-                ModelingToolkit.equations(dds),
-            )
-        else
-            throw(
-                error(
-                    "Measured quantities (output functions) were not provided and no outputs were found.",
-                ),
-            )
-        end
-    end
-
-    # Converting the finite difference operator in the right-hand side to
-    # the corresponding shift operator
-    eqs = filter(eq -> !(ModelingToolkit.isoutput(eq.lhs)), ModelingToolkit.equations(dds))
-    deltas = [Symbolics.operation(e.lhs).dt for e in eqs]
-    @assert length(Set(deltas)) == 1
-    eqs_shift = [e.lhs ~ e.rhs + first(Symbolics.arguments(e.lhs)) for e in eqs]
-    dds_shift = DiscreteSystem(eqs_shift, name = gensym())
-    @debug "System transformed from difference to shift: $dds_shift"
-
-    dds_aux, conversion = mtk_to_si(dds_shift, measured_quantities)
-    if length(funcs_to_check) == 0
-        params = parameters(dds)
-        params_from_measured_quantities = union(
-            [filter(s -> !istree(s), get_variables(y)) for y in measured_quantities]...,
-        )
-        funcs_to_check = vcat(
-            [x for x in states(dds) if conversion[x] in dds_aux.x_vars],
-            union(params, params_from_measured_quantities),
-        )
-    end
-    funcs_to_check_ = [eval_at_nemo(x, conversion) for x in funcs_to_check]
-    known_ic_ = [eval_at_nemo(x, conversion) for x in known_ic]
-
-    result = _assess_local_identifiability_discrete_aux(
-        dds_aux,
-        funcs_to_check_,
-        known_ic_,
-        prob_threshold,
-    )
-    nemo2mtk = Dict(funcs_to_check_ .=> funcs_to_check)
-    out_dict = OrderedDict(nemo2mtk[param] => result[param] for param in funcs_to_check_)
-    if length(known_ic) > 0
-        @warn "Since known initial conditions were provided, identifiability of states (e.g., `x(t)`) is at t = 0 only !"
-        out_dict = OrderedDict(substitute(k, Dict(t => 0)) => v for (k, v) in out_dict)
-    end
-    return out_dict
-end
-# ------------------------------------------------------------------------------
