+++ src/identifiable_functions.jl
@@ -117,93 +117,3 @@ function _find_identifiable_functions(
 
     return id_funcs_fracs
 end
-
-"""
-    find_identifiable_functions(ode::ModelingToolkit.ODESystem; measured_quantities=[], options...)
-
-Finds all functions of parameters/states that are identifiable in the given ODE
-system.
-
-## Options
-
-This functions takes the following optional arguments:
-- `measured_quantities` - the output functions of the model.
-- `loglevel` - the verbosity of the logging 
-  (can be Logging.Error, Logging.Warn, Logging.Info, Logging.Debug)
-
-## Example
-
-```jldoctest
-using StructuralIdentifiability
-using ModelingToolkit
-
-@parameters a01 a21 a12
-@variables t x0(t) x1(t) y1(t) [output = true]
-D = Differential(t)
-
-eqs = [
-    D(x0) ~ -(a01 + a21) * x0 + a12 * x1, 
-    D(x1) ~ a21 * x0 - a12 * x1, y1 ~ x0
-]
-de = ODESystem(eqs, t, name = :Test)
-
-find_identifiable_functions(de, measured_quantities = [y1 ~ x0])
-
-# prints
-2-element Vector{Num}:
-         a01*a12
- a01 + a12 + a21
-```
-"""
-function find_identifiable_functions(
-    ode::ModelingToolkit.ODESystem;
-    measured_quantities = Array{ModelingToolkit.Equation}[],
-    prob_threshold::Float64 = 0.99,
-    seed = 42,
-    with_states = false,
-    simplify = :standard,
-    rational_interpolator = :VanDerHoevenLecerf,
-    loglevel = Logging.Info,
-)
-    restart_logging(loglevel = loglevel)
-    reset_timings()
-    with_logger(_si_logger[]) do
-        return _find_identifiable_functions(
-            ode,
-            measured_quantities = measured_quantities,
-            prob_threshold = prob_threshold,
-            seed = seed,
-            with_states = with_states,
-            simplify = simplify,
-            rational_interpolator = rational_interpolator,
-        )
-    end
-end
-
-function _find_identifiable_functions(
-    ode::ModelingToolkit.ODESystem;
-    measured_quantities = Array{ModelingToolkit.Equation}[],
-    prob_threshold::Float64 = 0.99,
-    seed = 42,
-    with_states = false,
-    simplify = :standard,
-    rational_interpolator = :VanDerHoevenLecerf,
-)
-    Random.seed!(seed)
-    if isempty(measured_quantities)
-        measured_quantities = get_measured_quantities(ode)
-    end
-    ode, conversion = mtk_to_si(ode, measured_quantities)
-    result = _find_identifiable_functions(
-        ode,
-        simplify = simplify,
-        prob_threshold = prob_threshold,
-        seed = seed,
-        with_states = with_states,
-        rational_interpolator = rational_interpolator,
-    )
-    result = [parent_ring_change(f, ode.poly_ring) for f in result]
-    nemo2mtk = Dict(v => Num(k) for (k, v) in conversion)
-    out_funcs = [eval_at_dict(func, nemo2mtk) for func in result]
-    return out_funcs
-end
