+++ src/local_identifiability.jl
@@ -142,104 +142,6 @@ function get_degree_and_coeffsize(f::Generic.Frac{<:MPolyElem{Nemo.fmpq}})
 end
 
 # ------------------------------------------------------------------------------
-"""
-    function assess_local_identifiability(ode::ModelingToolkit.ODESystem; measured_quantities=Array{ModelingToolkit.Equation}[], funcs_to_check=Array{}[], prob_threshold::Float64=0.99, type=:SE, loglevel=Logging.Info)
-
-Input:
-- `ode` - the ODESystem object from ModelingToolkit
-- `measured_quantities` - the measurable outputs of the model
-- `funcs_to_check` - functions of parameters for which to check identifiability
-- `prob_threshold` - probability of correctness
-- `type` - identifiability type (`:SE` for single-experiment, `:ME` for multi-experiment)
-- `loglevel` - the minimal level of log messages to display (`Logging.Info` by default)
-
-Output:
-- for `type=:SE`, the result is an (ordered) dictionary from each parameter to boolean;
-- for `type=:ME`, the result is a tuple with the dictionary as in `:SE` case and array of number of experiments.
-
-The function determines local identifiability of parameters in `funcs_to_check` or all possible parameters if `funcs_to_check` is empty
-
-The result is correct with probability at least `prob_threshold`.
-
-`type` can be either `:SE` (single-experiment identifiability) or `:ME` (multi-experiment identifiability).
-The return value is a tuple consisting of the array of bools and the number of experiments to be performed.
-"""
-function assess_local_identifiability(
-    ode::ModelingToolkit.ODESystem;
-    measured_quantities = Array{ModelingToolkit.Equation}[],
-    funcs_to_check = Array{}[],
-    prob_threshold::Float64 = 0.99,
-    type = :SE,
-    loglevel = Logging.Info,
-)
-    restart_logging(loglevel = loglevel)
-    with_logger(_si_logger[]) do
-        return _assess_local_identifiability(
-            ode,
-            measured_quantities = measured_quantities,
-            funcs_to_check = funcs_to_check,
-            prob_threshold = prob_threshold,
-            type = type,
-        )
-    end
-end
-
-@timeit _to function _assess_local_identifiability(
-    ode::ModelingToolkit.ODESystem;
-    measured_quantities = Array{ModelingToolkit.Equation}[],
-    funcs_to_check = Array{}[],
-    prob_threshold::Float64 = 0.99,
-    type = :SE,
-)
-    if length(measured_quantities) == 0
-        if any(ModelingToolkit.isoutput(eq.lhs) for eq in ModelingToolkit.equations(ode))
-            @info "Measured quantities are not provided, trying to find the outputs in input ODE."
-            measured_quantities = filter(
-                eq -> (ModelingToolkit.isoutput(eq.lhs)),
-                ModelingToolkit.equations(ode),
-            )
-        else
-            throw(
-                error(
-                    "Measured quantities (output functions) were not provided and no outputs were found.",
-                ),
-            )
-        end
-    end
-    if length(funcs_to_check) == 0
-        funcs_to_check = vcat(
-            [e for e in ModelingToolkit.states(ode) if !ModelingToolkit.isoutput(e)],
-            ModelingToolkit.parameters(ode),
-        )
-    end
-    ode, conversion = mtk_to_si(ode, measured_quantities)
-    funcs_to_check_ = [eval_at_nemo(x, conversion) for x in funcs_to_check]
-
-    if isequal(type, :SE)
-        result = _assess_local_identifiability(
-            ode,
-            funcs_to_check = funcs_to_check_,
-            prob_threshold = prob_threshold,
-            type = type,
-        )
-        nemo2mtk = Dict(funcs_to_check_ .=> funcs_to_check)
-        out_dict =
-            OrderedDict(nemo2mtk[param] => result[param] for param in funcs_to_check_)
-        return out_dict
-    elseif isequal(type, :ME)
-        result, bd = _assess_local_identifiability(
-            ode,
-            funcs_to_check = funcs_to_check_,
-            prob_threshold = prob_threshold,
-            type = type,
-        )
-        nemo2mtk = Dict(funcs_to_check_ .=> funcs_to_check)
-        out_dict =
-            OrderedDict(nemo2mtk[param] => result[param] for param in funcs_to_check_)
-        return (out_dict, bd)
-    end
-end
-# ------------------------------------------------------------------------------
 
 """
     assess_local_identifiability(ode::ODE{P}; funcs_to_check::Array{<: Any, 1}, prob_threshold::Float64=0.99, type=:SE, loglevel=Logging.Info) where P <: MPolyElem{Nemo.fmpq}
